+++ src/components/Loading/index.js
@@ -1,3 +1,1 @@
-import Loading from 'carbon-components-react/lib/components/Loading';
-
-export { Loading as default };
+export { Loading } from 'carbon-components-react';
