+++ src/components/RadioTile/index.js
@@ -1,0 +1,1 @@
+export { RadioTile } from 'carbon-components-react';
