+++ src/components/Tile/index.js
@@ -1,0 +1,8 @@
+export {
+  Tile,
+  ClickableTile,
+  SelectableTile,
+  ExpandableTile,
+  TileAboveTheFoldContent,
+  TileBelowTheFoldContent,
+} from 'carbon-components-react';
